style: 홈페이지 꾸미기
* style: 네브바 수정 및 래퍼 수정

* refactor: 함수 재구성

index 파일도 수정

* refactor: 훅 폴더로 모바일 훅 이동

* style: 사진 다 추가

추가

---------
