+++ index.html
@@ -3,7 +3,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Vite + React + TS" />
+    <meta name="description" content="홍익대학교 컴퓨터공학과 밴드학회 큰소리입니다!" />
     <title>큰소리</title>
   </head>
   <body>
