+++ src/components/Wrapper.tsx
@@ -7,6 +7,7 @@ const Wrapper = styled.div`
   background: #fffbf3;
   margin-top: 50px;
   width: 100%;
+  overflow-x: hidden;
   height: 100%;
 `;
