+++ src/components/navBar/navBar.tsx
@@ -3,7 +3,7 @@ import { css } from "@emotion/css";
 import React, { useEffect, useState } from "react";
 import logo from "/image/logo.svg";
 import { Link, useLocation } from "react-router-dom";
-import useIsMobile from "../mobile/useIsMobile";
+import useIsMobile from "../../hooks/useIsMobile.tsx";
 import { IoClose, IoMenu } from "react-icons/io5";
 import Space from "./Space.tsx";
 import { Menu, MobileMenu } from "./Menu.tsx";
@@ -72,17 +72,13 @@ const NavBar: React.FC = () => {
           </Button>
           <MobileMenu isOpened={isOpen} isSmall={true}>
             <Link to="/recruit">
-              <Space isActive={location.pathname === "/recruit"}>
-                지원하기
-              </Space>
+              <Space isActive={false}>지원하기</Space>
             </Link>
             <Link to="/contact">
-              <Space isActive={location.pathname === "/contact"}>
-                문의하기
-              </Space>
+              <Space isActive={false}>문의하기</Space>
             </Link>
             <Link to="/login">
-              <Space isActive={location.pathname === "/login"}>my keun</Space>
+              <Space isActive={false}>my keun</Space>
             </Link>
           </MobileMenu>
         </>
