+++ src/components/navBar/navBar2.tsx
@@ -3,7 +3,7 @@ import { css } from "@emotion/css";
 import React, { useEffect, useState, useContext } from "react";
 import logo from "/image/logo.svg";
 import { Link, useLocation } from "react-router-dom";
-import useIsMobile from "../mobile/useIsMobile";
+import useIsMobile from "../../hooks/useIsMobile.tsx";
 import { IoClose, IoMenu } from "react-icons/io5";
 import Space from "./Space.tsx";
 import { Menu, MobileMenu } from "./Menu.tsx";
